started refactor of nav and routing to use flux
