+++ react-native-app/.eslintrc.js
@@ -1,8 +1,10 @@
+
 module.exports = {
     "extends": "airbnb",
     "rules": {
         "react/jsx-filename-extension": "off",
-        "react/no-unused-state": "off"
+        "react/no-unused-state": "off",
+        "react/prefer-stateless-function": "off"
     }
 
 };
