+++ react-native-app/package.json
@@ -25,6 +25,7 @@
   "dependencies": {
     "axios": "^0.17.1",
     "eslint": "^4.11.0",
+    "eslint-config-airbnb": "^16.1.0",
     "expo": "^22.0.2",
     "react": "16.0.0-beta.5",
     "react-addons-update": "^15.6.2",
@@ -32,7 +33,7 @@
     "react-native-elements": "^0.18.2",
     "react-native-map": "0.0.1",
     "react-native-maps": "*",
-    "react-native-router-flux": "^4.0.0-beta.24",
+    "react-native-router-flux": "^3.35.0",
     "react-navigation": "^1.0.0-beta.19",
     "react-redux": "^5.0.6",
     "redux": "^3.7.2",
